:truck: Moved the test location
