+++ src/test/java/faker/generator/RobinTest.java
@@ -1,4 +1,4 @@
-package faker.test;
+package faker.generator;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
